+++ langsmith_migrator/core/migrators/__init__.py
@@ -3,6 +3,7 @@
 from .base import BaseMigrator
 from .dataset import DatasetMigrator
 from .experiment import ExperimentMigrator
+from .feedback import FeedbackMigrator
 from .annotation_queue import AnnotationQueueMigrator
 from .prompt import PromptMigrator
 from .rules import RulesMigrator
@@ -13,6 +14,7 @@ __all__ = [
     "BaseMigrator",
     "DatasetMigrator",
     "ExperimentMigrator",
+    "FeedbackMigrator",
     "AnnotationQueueMigrator",
     "PromptMigrator",
     "RulesMigrator",